+++ GeosCore/drydep_mod.F90
@@ -1007,6 +1007,7 @@ CONTAINS
     USE Input_Opt_Mod,      ONLY : OptInput
     USE Species_Mod,        ONLY : Species
     USE State_Chm_Mod,      ONLY : ChmState
+    USE State_Chm_Mod,      ONLY : Ind_
     USE State_Diag_Mod,     ONLY : DgnState
     USE State_Grid_Mod,     ONLY : GrdState
     USE State_Met_Mod,      ONLY : MetState
@@ -1226,6 +1227,10 @@ CONTAINS
     ! Small number
     REAL(fp), PARAMETER :: SMALL = 1.0e-10_f8
 
+    ! Gravitational Settling velocity (yli, 3/28/2024)
+    REAL(f8)  :: VTSoutput ! Used in AERO_SFCRSII, ADUST_SFCRSII and DUST_SFCRSII for a specific aerosol species and a specific land type
+    REAL(f8)  :: VTSoutput_ (NUMDEP,NTYPE) ! For all deposited aerosol species and all land types
+
     !=================================================================
     ! DEPVEL begins here!
     !=================================================================
@@ -1339,11 +1344,11 @@ CONTAINS
     !$OMP PRIVATE( DUMMY1,    DUMMY2,  DUMMY3,  DUMMY4, DAIR,    RB         )&
     !$OMP PRIVATE( C1X,       VK,      I,       J,      IW                  )&
     !$OMP PRIVATE( DIAM,      DEN,     XLAI_FP, SUNCOS_FP,       CFRAC_FP   )&
-    !$OMP PRIVATE( N_SPC,     alpha,   DEPVw                                )&
+    !$OMP PRIVATE( N_SPC,     alpha,   DEPVw    VTSoutput,       VTSoutput_ )&
 #ifdef TOMAS
     !$OMP PRIVATE( BIN                                                      )&
 #endif
-    !$OMP PRIVATE( SpcId,  SpcInfo,    F0_K                                 )&
+    !$OMP PRIVATE( SpcId,   SpcInfo,    F0_K                                )&
     !$OMP COLLAPSE( 2                                                       )
     DO J = 1, State_Grid%NY
     DO I = 1, State_Grid%NX
@@ -1402,6 +1407,8 @@ CONTAINS
        alpha      = 0.0_f8
        DEPVw      = 0.0_f8
        F0_K       = 0.0_f8
+       VTSoutput  = 0.0_fp
+       VTSoutput_ = 0.0_fp
 
        !** CZ is Altitude (m) at which deposition velocity is computed
        CZ    = CZ1(I,J)
@@ -1721,7 +1728,11 @@ CONTAINS
                                               USTAR(I,J),          &
                                               RHB(I,J),            &
                                               W10(I,J),            &
-                                              Input_Opt )
+                                              VTSoutput,           &
+                                              Input_Opt,           & 
+                                              State_Chm )
+
+                VTSoutput_(K,LDT) = VTSoutput  
 
              ELSE IF ( SpcInfo%DD_DustDryDep ) THEN
 
@@ -1765,6 +1776,34 @@ CONTAINS
                 ! Particle diameter, convert [m] -> [um]
                 DIAM  = A_RADI(K) * 2.e+0_f8
 
+                IF ((Ind_(SpcInfo%Name) .EQ. Ind_('DST1')) .OR.      &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('DSTAL1')) .OR.    &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('NITD1')) .OR.     &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('SO4D1'))) THEN
+                    DIAM = 0.66895E-6
+                ENDIF
+
+                IF ((Ind_(SpcInfo%Name) .EQ. Ind_('DST2')) .OR.      &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('DSTAL2')) .OR.    &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('NITD2')) .OR.     &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('SO4D2'))) THEN
+                    DIAM = 2.4907E-6
+                ENDIF
+
+                IF ((Ind_(SpcInfo%Name) .EQ. Ind_('DST3')) .OR.      &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('DSTAL3')) .OR.    &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('NITD3')) .OR.     &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('SO4D3'))) THEN
+                    DIAM = 4.164E-6
+                ENDIF
+
+                IF ((Ind_(SpcInfo%Name) .EQ. Ind_('DST4')) .OR.      &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('DSTAL4')) .OR.    &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('NITD4')) .OR.     &
+                    (Ind_(SpcInfo%Name) .EQ. Ind_('SO4D4'))) THEN
+                    DIAM = 6.677E-6
+                ENDIF
+
                 ! Particle density [kg/m3]
                 DEN   = A_DEN(K)
 #endif
@@ -1805,7 +1844,10 @@ CONTAINS
                                               TEMPK,               &
                                               USTAR(I,J),          &
                                               DIAM,                &
-                                              DEN )
+                                              DEN,                 &
+                                              VTSoutput )
+
+                VTSoutput_(K,LDT) = VTSoutput
 
              ELSE
 
@@ -1828,7 +1870,11 @@ CONTAINS
                 !   (-CZH)**0.6667D0
 
                 RSURFC(K, LDT) = ADUST_SFCRSII(K, II, PRESSU(I,J)*1e-3_f8, &
-                                            TEMPK, USTAR(I,J))
+                                            TEMPK, USTAR(I,J), &
+                                            VTSoutput, RHB(I,J), &
+                                            State_Chm)
+
+                VTSoutput_(K,LDT) = VTSoutput
 
                 !*
                 !*    Set VDS to be less than VDSMAX (entry in input file
@@ -2183,6 +2229,24 @@ CONTAINS
              IF (.NOT.LDEP(K)) GOTO 400
              VK(K) = VD(K)
              VD(K) = VK(K) +.001D0* DBLE( IUSE(I,J,LDT) )/C1X(K)
+
+             IF (AIROSOL (K)) THEN
+                 SpcId   =  NTRAIND(K)
+                 SpcInfo => State_Chm%SpcData(SpcId)%Info
+
+                 IF ( SpcInfo%DD_AeroDryDep ) THEN
+                     VD(K) = VK(K) +.001D0* DBLE( IUSE(I,J,LDT) )  &
+       /C1X(K) + .001D0* DBLE( IUSE(I,J,LDT) ) *VTSoutput_(K,LDT)
+                 ELSE IF ( SpcInfo%DD_DustDryDep ) THEN
+                     VD(K) = VK(K) +.001D0* DBLE( IUSE(I,J,LDT) )  &
+       /C1X(K) + .001D0* DBLE( IUSE(I,J,LDT) ) *VTSoutput_(K,LDT)
+                 ELSE
+                     VD(K) = VK(K) +.001D0* DBLE( IUSE(I,J,LDT) )  &
+       /C1X(K) + .001D0* DBLE( IUSE(I,J,LDT) ) *VTSoutput_(K,LDT)
+                 ENDIF
+             ELSE
+                 VD(K) = VK(K) +.001D0* DBLE( IUSE(I,J,LDT) )/C1X(K)
+             ENDIF
 400       CONTINUE
 500    CONTINUE
 
@@ -3130,12 +3194,15 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  FUNCTION AERO_SFCRSII( K, II, PRESS, TEMP, USTAR, RHB, W10, Input_Opt ) &
+  FUNCTION AERO_SFCRSII( K, II, PRESS, TEMP, USTAR, RHB, W10, VTSout, Input_Opt, State_Chm ) &
        RESULT(RS)
 !
 ! !USES:
 !
     USE Input_Opt_Mod,      ONLY : OptInput
+    USE Species_Mod,        ONLY : Species
+    USE State_Chm_Mod,      ONLY : ChmState
+    USE State_Chm_Mod,      ONLY : Ind_
 !
 ! !INPUT PARAMETERS:
 !
@@ -3147,6 +3214,13 @@ CONTAINS
     REAL(f8),       INTENT(IN) :: RHB   ! Relative humidity (fraction)
     REAL(f8),       INTENT(IN) :: W10   ! 10 m windspeed [m/s]
     TYPE(OptInput), INTENT(IN) :: Input_Opt ! Input Options object
+    TYPE(ChmState), INTENT(IN) :: State_Chm   ! Chemistry State object
+
+!
+! !OUTPUT PARAMETERS: 
+!
+    REAL(f8),       INTENT(OUT) :: VTSout    ! Settling velocity [m/s]
+
 !
 ! !RETURN VALUE:
 !
@@ -3229,6 +3303,11 @@ CONTAINS
     REAL(f8)      :: DEDGE
     REAL(f8)      :: DEN1, WTP
     INTEGER       :: ID,NR
+
+    ! For the species database
+    INTEGER                :: SpcId
+    TYPE(Species), POINTER :: SpcInfo
+    REAL(f8)               :: drydepRadius
 
     !=======================================================================
     !   #  LUC [Zhang et al., 2001]                GEOS-CHEM LUC (Corr. #)
@@ -3403,11 +3482,29 @@ CONTAINS
     NR = INT((( Input_Opt%SALC_REDGE_um(2) - Input_Opt%SALA_REDGE_um(1) ) &
          / DR ) + 0.5e+0_f8 )
 
+    ! drydepRadius
+    SpcId = NTRAIND(K)
+    SpcInfo => State_Chm%SpcData(SpcId)%Info
+
+    drydepRadius = A_RADI(K)
+
+    ! Coarse seasalt
+    IF ((Ind_(SpcInfo%Name) .EQ. Ind_('NITS')) .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('SALC')) & 
+  .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('SO4S')) .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('BRSALC')) &
+  .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('ISALC'))) THEN
+           drydepRadius = 0.74025E-6
+    ENDIF
+
+    IF ((Ind_(SpcInfo%Name) .EQ. Ind_('SALA')) .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('BRSALA')) &
+        .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('ISALA'))) THEN
+           drydepRadius = 0.114945E-6
+    ENDIF
+
     ! Particle radius [cm]
     ! Bug fix: The Gerber [1985] growth should use the dry radius
     ! in micromenters and not cm. Replace RCM with RUM (jaegle 5/11/11)
     !RCM  = A_RADI(K) * 1.d2
-    RUM  = A_RADI(K) * 1.e+6_f8
+    RUM  = drydepRadius * 1.e+6_f8
 
     ! Exponential factors used for hygroscopic growth
     ! Replace RCM with RUM (jaegle 5/11/11)
@@ -3433,7 +3530,7 @@ CONTAINS
 
     ! Use equation 5 in Lewis and Schwartz (2006) for sea salt growth [m]
     ! (jaegle 5/11/11)
-    RWET = A_RADI(K) * (4.e+0_f8 / 3.7e+0_f8) * &
+    RWET = drydepRadius * (4.e+0_f8 / 3.7e+0_f8) * &
           ( (2.e+0_f8 - RHBL)/(1.e+0_f8 - RHBL) )**(1.e+0_f8/3.e+0_f8)
 
     ! Ratio dry over wet radii at the cubic power
@@ -3451,7 +3548,7 @@ CONTAINS
     ! wet aerosol (kg/m3)
     ! (bec, 6/17/10, jaegle 5/11/11)
     ! Redefine RATIO_R
-    RATIO_R = A_RADI(K) / RWET
+    RATIO_R = drydepRadius / RWET
 
     ! Assume an initial density of 1000 kg/m3
     DEN  = 1000.e+0_f8
@@ -3531,19 +3628,20 @@ CONTAINS
        ! Calculate mass of wet aerosol (Dw = wet diameter, D = dry diamter):
        ! Overall = dM/dDw = dV/dlnD * Rwet/Rdry * DEN /Rw
        IF (DMID(ID) .ge. D0 .and. DMID(ID) .le. D1 ) THEN
-          DMIDW = DMID(ID) * RWET/A_RADI(K)   ! wet radius [um]
-          SALT_MASS   = SALT_V(ID) * RWET/A_RADI(K) * DEN / &
+          DMIDW = DMID(ID) * RWET/drydepRadius   ! wet radius [um]
+          SALT_MASS   = SALT_V(ID) * RWET/drydepRadius * DEN / &
                         (DMIDW*0.5e+0_f8)
           VTS_WEIGHT  = VTS_WEIGHT + &
                         SALT_MASS * VTS * (DMIDW/(RWET*1d6*2e+0_f8) )** &
-                        2e+0_f8 * (2e+0_f8 * DR *  RWET/A_RADI(K))
+                        2e+0_f8 * (2e+0_f8 * DR *  RWET/drydepRadius)
           SALT_MASS_TOTAL = SALT_MASS_TOTAL+SALT_MASS * &
-                            (2e+0_f8 * DR *  RWET/A_RADI(K))
+                            (2e+0_f8 * DR *  RWET/drydepRadius)
        ENDIF
     ENDDO
 
     ! Final mass weighted setting velocity:
     VTS = VTS_WEIGHT/SALT_MASS_TOTAL
+    VTSout = VTS
 
     ! Brownian diffusion constant for particle (m2/s)
     DIFF = BOLTZ * TEMP * SLIP / (3.e+0_f8 * PI * VISC * DIAM)
@@ -3893,7 +3991,15 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  FUNCTION ADUST_SFCRSII( K, II, PRESS, TEMP, USTAR ) RESULT( RS )
+  FUNCTION ADUST_SFCRSII( K, II, PRESS, TEMP, USTAR, &
+                          VTSout, RHB, State_Chm ) RESULT( RS )
+!
+! !USES: 
+!
+      USE Species_Mod,        ONLY : Species
+      USE State_Chm_Mod,      ONLY : ChmState
+      USE State_Chm_Mod,      ONLY : Ind_
+
 !
 ! !INPUT PARAMETERS:
 !
@@ -3902,6 +4008,14 @@ CONTAINS
     REAL(f8), INTENT(IN) :: PRESS ! Pressure [kPa] (1 mb = 100 Pa = 0.1 kPa)
     REAL(f8), INTENT(IN) :: TEMP  ! Temperature [K]
     REAL(f8), INTENT(IN) :: USTAR ! Friction velocity [m/s]
+    REAL(f8), INTENT(IN) :: RHB   ! Relative humidity (fraction)
+    TYPE(ChmState), INTENT(IN) :: State_Chm   ! Chemistry State object
+
+!
+! !OUTPUT PARAMETERS: 
+!
+    REAL(f8), INTENT(OUT) :: VTSout ! Settling velocity [m/s]
+
 !
 ! !RETURN VALUE:
 !
@@ -3944,6 +4058,12 @@ CONTAINS
     REAL(f8)  :: DIAM, DEN
 
     REAL(f8)  :: EB, EIM, EIN, R1, AA, VTS
+
+    REAL(f8)  :: RHBL        ! Relative humidity local
+
+    ! For the species database
+    INTEGER                :: SpcId
+    TYPE(Species), POINTER :: SpcInfo
 
     !=======================================================================
     !   #  LUC [Zhang et al., 2001]                GEOS-CHEM LUC (Corr. #)
@@ -4080,7 +4200,52 @@ CONTAINS
     !=================================================================
 
     ! Particle diameter [m] hotp 10/26/07
-    DIAM  = 0.5e-6_f8
+    DIAM  = 0.17378e-6_f8
+
+    ! Hygroscopic growth following Latimer and Martin (2019) ACP (yanshun)
+    ! Added safety check for LOG (phs, 6/11/08)
+    RHBL    = MAX( TINY(RHB), RHB )
+
+    ! Over oceans the RH in the viscous sublayer is set to 98%,
+    ! following
+    ! Lewis and Schwartz (2004), see discussion above (jaegle 5/11/11)
+    IF (LUC == 14) THEN
+        RHBL = 0.98
+    ENDIF
+
+    SpcId = NTRAIND(K)
+    SpcInfo => State_Chm%SpcData(SpcId)%Info
+
+    ! SIA
+    IF ((Ind_(SpcInfo%Name) .EQ. Ind_('NIT')) .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('NH4')) &
+        .OR. (Ind_(SpcInfo%Name) .EQ. Ind_('SO4'))) THEN
+        ! Efflorescence transtions
+        IF (RHBL .LT. 0.35) THEN
+            DIAM = DIAM * 1.0
+        ! Linear hygroscopic growth    
+        ELSE IF ((RHBL .GE. 0.35) .AND. (RHBL .LE. 0.40)) THEN
+            DIAM = DIAM + (DIAM * ((1.0 + 0.61 * 0.40 / &
+                   (1.0 - 0.40)) ** (1.0 / 3.0)) - DIAM) / &
+                   (0.40 - 0.35) * (RHBL - 0.35)
+        ! Kohler hygroscopic growth
+        ELSE
+            DIAM = DIAM * ((1.0 + 0.61 * RHBL / (1.0 - RHBL)) &
+                   ** (1.0 / 3.0))
+        ENDIF
+
+    !BC
+    ELSE IF ((Ind_(SpcInfo%Name) .EQ. Ind_('BCPI')) .OR. &
+       (Ind_(SpcInfo%Name) .EQ. Ind_('BCPO'))) THEN
+        DIAM = DIAM * 1.0
+    
+    !OA
+    ELSE
+        DIAM = DIAM * ((1.0 + 0.1 * RHBL / (1.0 - RHBL)) &
+      ** (1.0 / 3.0))
+    ENDIF
+
+    ! Free pointer
+    SpcInfo => NULL()
 
     ! Particle density [kg/m3] hotp 10/26/07
     DEN   = 1500
@@ -4124,6 +4289,7 @@ CONTAINS
 
     ! Settling velocity [m/s]
     VTS  = CONST * SLIP / VISC
+    VTSout = VTS
 
     ! Brownian diffusion constant for particle (m2/s)
     DIFF = BOLTZ * TEMP * SLIP / (3.e+0_f8 * Pi * VISC * DIAM)
@@ -4182,8 +4348,8 @@ CONTAINS
 ! !INTERFACE:
 !
 
-  FUNCTION DUST_SFCRSII( K, II, PRESS, TEMP, USTAR, DIAM, DEN ) &
-       RESULT( RS )
+  FUNCTION DUST_SFCRSII( K, II, PRESS, TEMP, USTAR, DIAM, &
+                         DEN, VTSout) RESULT( RS )
 !
 ! !INPUT PARAMETERS:
 !
@@ -4194,6 +4360,12 @@ CONTAINS
     REAL(f8), INTENT(IN) :: USTAR   ! Friction velocity [m/s]
     REAL(f8), INTENT(IN) :: DIAM    ! Particle diameter [m]
     REAL(f8), INTENT(IN) :: DEN     ! Particle density [kg/m3]
+
+!
+! !OUTPUT PARAMETERS: 
+!
+    REAL(f8), INTENT(OUT) :: VTSout ! Settling velocity [m/s]
+
 !
 ! !RETURN VALUE:
 !
@@ -4409,6 +4581,7 @@ CONTAINS
 
     ! Settling velocity [m/s]
     VTS  = CONST * SLIP / VISC
+    VTSout = VTS
 
     ! Brownian diffusion constant for particle (m2/s)
     DIFF = BOLTZ * TEMP * SLIP / (3.e+0_f8 * Pi * VISC * DIAM)